ai-pm-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
redis==5.2.1
pydantic==2.10.3
python-multipart==0.0.19
python-dotenv==1.0.1
httpx==0.28.1
google-generativeai==0.8.3
python-socketio==5.12.0
aiofiles==23.2.1
jinja2==3.1.2
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
websockets==12.0
aiohttp==3.9.1
python-engineio==4.7.1
pytest==7.4.3
pytest-asyncio==0.21.1